Make apply method an external file, make register*Plugins external files and not methods
